Split ble.d.ts
