+++ examples/manifest_typings.json
@@ -4,7 +4,7 @@
 	},
 	"modules": {
 		"*": [
-			"$(TYPINGS)/*"
+			"$(TYPINGS)/*", "$(TYPINGS)/ble/*"
 		],
 		"commodetto/*": [
 			"$(TYPINGS)/commodetto/*"
